Extract common code
