+++ plugins/convention/src/main/java/org/apache/struts2/convention/DefaultActionNameBuilder.java
@@ -21,11 +21,6 @@
 package org.apache.struts2.convention;
 
 import com.opensymphony.xwork2.inject.Inject;
-import com.opensymphony.xwork2.util.TextParseUtil;
-import org.apache.commons.lang3.StringUtils;
-
-import java.util.Collections;
-import java.util.Set;
 
 /**
  * <p>
@@ -36,28 +31,21 @@ import java.util.Set;
  * action names.
  * </p>
  */
-public class DefaultActionNameBuilder implements ActionNameBuilder {
-    private Set<String> actionSuffix = Collections.singleton("Action");
+public class DefaultActionNameBuilder extends AbstractActionNameBuilder {
+
     private boolean lowerCase;
 
     @Inject
-    public DefaultActionNameBuilder(@Inject(value="struts.convention.action.name.lowercase") String lowerCase) {
+    public DefaultActionNameBuilder(
+            @Inject(ConventionConstants.CONVENTION_ACTION_NAME_LOWERCASE) String lowerCase
+    ) {
         this.lowerCase = Boolean.parseBoolean(lowerCase);
-    }
-
-    /**
-     * @param   actionSuffix (Optional) Classes that end with these value will be mapped as actions
-     *          (defaults to "Action")
-     */
-    @Inject(value = "struts.convention.action.suffix", required = false)
-    public void setActionSuffix(String actionSuffix) {
-        if (StringUtils.isNotBlank(actionSuffix)) {
-            this.actionSuffix = TextParseUtil.commaDelimitedStringToSet(actionSuffix);
-        }
     }
 
     public String build(String className) {
         String actionName = className;
+
+        checkActionName(actionName);
 
         // Truncate Action suffix if found
         actionName = truncateSuffixIfMatches(actionName);
@@ -75,13 +63,4 @@ public class DefaultActionNameBuilder implements ActionNameBuilder {
         return actionName;
     }
 
-    private String truncateSuffixIfMatches(String name) {
-        String actionName = name;
-        for (String suffix : actionSuffix) {
-            if (actionName.endsWith(suffix)) {
-                actionName = actionName.substring(0, actionName.length() - suffix.length());
-            }
-        }
-        return actionName;
-    }
 }
